have HomeLocation extend Location
